requirements.txt
pdfplumber~=0.11.8
pandas~=2.3.3
python-dateutil~=2.9.0.post0
pillow~=12.0.0
easyocr
pdf2image